Move the exercise details page
- Move view file
- Reconfigure router
